UPSTREAM: Update degree symbol to utf-8 encoding in comments
Almost all of the places where we have the degree symbol '', it's
encoded as 0xc2 0xb0 (utf-8 encoding).  There are a few places where it
is encoded as just a high ascii byte: 0xb0.  Editors that support the
high ascii 0xb0 seem to support the utf-8 0xc2 0xb0 encoding as well,
but the opposite does not seem to be true.

Change the high-ascii degree symbols to utf-8 encoding.

BUG=None
BRANCH=None
TEST=None

Change-Id: I4b2fc603d57ab140c6bd6a9e8fe3eaa480d39d68
Signed-off-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://review.coreboot.org/15978
Tested-by: build bot (Jenkins)
Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-by: Alexander Couzens <lynxis@fe80.eu>
Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Reviewed-on: https://chromium-review.googlesource.com/366219
Commit-Ready: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>