Revert "gru: Show the current time on start-up"
This reverts commit d0361193e0ec135e21f0611d7fa6e5c02f2b2bfc.

google_chromeec_init() is a weird function that can lead to confusing
behavior. I'm not sure how it's meant to work on the boards that use it,
but it causes problems on Kevin and other non-x86 boards have never used
it either. It doesn't really do anything anyway (the EC works fine
without an initial HELLO), so at best it's just a waste of time... let's
take it back out.

There's also no need to display the current time on every boot... other
boards don't do that and the eventlog already fills the same purpose.
Cut it out to avoid one extra host command overhead.

BRANCH=None
BUG=chrome-os-partner:55995
TEST=Recovery reasons now get correctly propagated across the EC reboot.

Change-Id: I58fd5e6094e1c8cb6368e7a4569ab9231375fbc9
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/367351
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Shelley Chen <shchen@chromium.org>